refactor action.h dependency direction and normalize scores
